+++ warp/payload/payload_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package messages
+package payload
 
 import (
 	"testing"
